requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP clients
httpx>=0.26.0
aiohttp>=3.9.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image processing
Pillow>=10.2.0

# OpenAI (DALL-E 3)
openai>=1.10.0

# Replicate (Flux, SDXL)
replicate>=0.22.0

# Stability AI
stability-sdk>=0.8.0

# Storage
boto3>=1.34.0  # For S3-compatible storage (optional)

# Environment management
python-dotenv>=1.0.0

# Utilities
aiofiles>=23.2.0
